Separate constants for PHY and MAC. Refactor user-configurable constants
- Move CSMA constants to MAC
- Add constant module to PHY
- Remove 'user_configurable_constants' module. Now included in MAC constants
- Update 'build.rs' script
